Safari-611.1.10/LayoutTests/fast/text-autosizing/ios/idempotentmode/line-height-boosting-expected.txt

PASS result is >= 18
PASS result is >= 13
PASS result is >= 11
PASS result is 12
PASS result is NaN
PASS result is 18
PASS result is 18
PASS result is 18
PASS successfullyParsed is true

TEST COMPLETE
yes line-height boost
yes line-height boost
yes line-height boost
no line-height boost
no line-height boost
no line-height boost
no line-height boost
no line-height boost
